tweetme/requirements.txt
argon2-cffi==18.1.0
awsebcli==3.14.3
bcrypt==3.1.4
boto3==1.7.65
botocore==1.10.65
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
Django==1.10
django-bootstrap3==10.0.1
django-common-helpers==0.9.1
django-crispy-forms==1.7.2
django-cron==0.5.1
django-crontab==0.7.1
django-filter==2.0.0
djangorestframework==3.8.2
docutils==0.14
Faker==0.8.16
idna==2.7
jmespath==0.9.3
Markdown==2.6.11
pathspec==0.5.5
Pillow==5.2.0
pycparser==2.18
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
stripe==2.1.0
tabulate==0.7.5
termcolor==1.1.0
text-unidecode==1.2
urllib3==1.23
wincertstore==0.2
